+++ GreedyParameterPointPreDefinedSampler.h
@@ -52,7 +52,7 @@ public:
                                    the different parameter points.
      * @param[in] check_local_rom Compute local ROM residual each iteration.
      * @param[in] tolerance A tolerance value for which to end the algorithm.
-     * @param[in] saturation A saturation constant.
+     * @param[in] alpha A alpha constant.
      * @param[in] subset_size The size of the random subset.
      * @param[in] convergence_subset_size The size of the convergence subset.
      * @param[in] use_centroid Whether to use the centroid heuristic when
@@ -65,7 +65,7 @@ public:
         std::vector<Vector> parameter_points,
         bool check_local_rom,
         double relative_error_tolerance,
-        double saturation,
+        double alpha,
         int subset_size,
         int convergence_subset_size,
         std::string output_log_path = "",
@@ -78,7 +78,7 @@ public:
         std::vector<double> parameter_points,
         bool check_local_rom,
         double relative_error_tolerance,
-        double saturation,
+        double alpha,
         int subset_size,
         int convergence_subset_size,
         std::string output_log_path = "",
